refactor(fromEvent): make fromEvent just a function - updates tests as well
