+++ spec/observables/fromEvent-spec.ts
@@ -1,28 +1,27 @@
 import { expect } from 'chai';
-import * as Rx from '../../src/Rx';
 import { expectObservable } from '../helpers/marble-testing';
+import { fromEvent, never, timer } from '../../src';
+import { TestScheduler } from '../../src/testing';
 
 declare function asDiagram(arg: string): Function;
-declare const rxTestScheduler: Rx.TestScheduler;
-
-const Observable = Rx.Observable;
+declare const rxTestScheduler: TestScheduler;
 
 /** @test {fromEvent} */
-describe('Observable.fromEvent', () => {
+describe('fromEvent', () => {
   asDiagram('fromEvent(element, \'click\')')
   ('should create an observable of click on the element', () => {
     const target = {
       addEventListener: (eventType, listener) => {
-        Observable.timer(50, 20, rxTestScheduler)
+        timer(50, 20, rxTestScheduler)
           .mapTo('ev')
           .take(2)
-          .concat(Observable.never())
+          .concat(never())
           .subscribe(listener);
       },
       removeEventListener: () => void 0,
       dispatchEvent: () => void 0,
     };
-    const e1 = Observable.fromEvent(target, 'click');
+    const e1 = fromEvent(target, 'click');
     const expected = '-----x-x---';
     expectObservable(e1).toBe(expected, {x: 'ev'});
   });
@@ -44,7 +43,7 @@ describe('Observable.fromEvent', () => {
       }
     };
 
-    const subscription = Observable.fromEvent(obj, 'click')
+    const subscription = fromEvent(obj, 'click')
       .subscribe(() => {
         //noop
        });
@@ -74,7 +73,7 @@ describe('Observable.fromEvent', () => {
       }
     };
 
-    const subscription = Observable.fromEvent(<any>obj, 'click')
+    const subscription = fromEvent(<any>obj, 'click')
       .subscribe(() => {
         //noop
        });
@@ -104,7 +103,7 @@ describe('Observable.fromEvent', () => {
       }
     };
 
-    const subscription = Observable.fromEvent(obj, 'click')
+    const subscription = fromEvent(obj, 'click')
       .subscribe(() => {
         //noop
        });
@@ -124,7 +123,7 @@ describe('Observable.fromEvent', () => {
       }
     };
 
-    Observable.fromEvent(obj as any, 'click').subscribe({
+    fromEvent(obj as any, 'click').subscribe({
       error(err) {
         expect(err).to.exist
           .and.be.instanceof(Error)
@@ -146,7 +145,7 @@ describe('Observable.fromEvent', () => {
       }
     };
 
-    const subscription = Observable.fromEvent(<any>obj, 'click', expectedOptions)
+    const subscription = fromEvent(<any>obj, 'click', expectedOptions)
       .subscribe(() => {
         //noop
        });
@@ -167,7 +166,7 @@ describe('Observable.fromEvent', () => {
       }
     };
 
-    Observable.fromEvent(obj, 'click').take(1)
+    fromEvent(obj, 'click').take(1)
       .subscribe((e: any) => {
         expect(e).to.equal('test');
       }, (err: any) => {
@@ -194,7 +193,7 @@ describe('Observable.fromEvent', () => {
       return x + '!';
     }
 
-    Observable.fromEvent(obj, 'click', selector).take(1)
+    fromEvent(obj, 'click', selector).take(1)
       .subscribe((e: any) => {
         expect(e).to.equal('test!');
       }, (err: any) => {
@@ -221,7 +220,7 @@ describe('Observable.fromEvent', () => {
       //noop
     }
 
-    Observable.fromEvent(obj, 'click', selector).take(1)
+    fromEvent(obj, 'click', selector).take(1)
       .subscribe((e: any) => {
         expect(e).not.exist;
       }, (err: any) => {
@@ -248,7 +247,7 @@ describe('Observable.fromEvent', () => {
       return 'no arguments';
     }
 
-    Observable.fromEvent(obj, 'click', selector).take(1)
+    fromEvent(obj, 'click', selector).take(1)
       .subscribe((e: any) => {
         expect(e).to.equal('no arguments');
       }, (err: any) => {
@@ -275,7 +274,7 @@ describe('Observable.fromEvent', () => {
       return [].slice.call(arguments);
     }
 
-    Observable.fromEvent(obj, 'click', selector).take(1)
+    fromEvent(obj, 'click', selector).take(1)
       .subscribe((e: any) => {
         expect(e).to.deep.equal([1, 2, 3]);
       }, (err: any) => {
@@ -298,7 +297,7 @@ describe('Observable.fromEvent', () => {
     const obj: NullProtoEventTarget = new NullProtoEventTarget();
 
     expect(() => {
-      Observable.fromEvent(obj, 'foo').subscribe();
+      fromEvent(obj, 'foo').subscribe();
       done();
     }).to.not.throw(TypeError);
   });
